+++ src/backend/joanie/core/models/accounts.py
@@ -28,11 +28,12 @@ class Address(models.Model):
     uid = models.UUIDField(
         default=uuid.uuid4, unique=True, editable=False, db_index=True
     )
-    name = models.CharField(_("name"), max_length=100)
+    title = models.CharField(_("title"), max_length=100)
     address = models.CharField(_("address"), max_length=255)
     postcode = models.CharField(_("postcode"), max_length=50)
     city = models.CharField(_("city"), max_length=255)
     country = CountryField(_("country"))
+    fullname = models.CharField(_("full name"), max_length=255)
     owner = models.ForeignKey(
         User,
         verbose_name=_("owner"),